chore: move active24 client into dns/internal
