+++ providers/dns/internal/active24/types.go
@@ -1,4 +1,4 @@
-package internal
+package active24
 
 import "fmt"
